Refactor: Add traits, consolidate constants, fix type consistency
- Add BearingCalculator trait, removing manual enum dispatch in main.rs
- Add NorthTracker trait for consistent north tracking interface
- Remove dead_code attributes from Filter trait (used by examples)
- Consolidate MIN_POWER_THRESHOLD constant to bearing.rs
- Standardize sample_rate to f32 across all signal processing
- Derive clap::ValueEnum directly on config enums, removing duplicates
- Update outdated accuracy comments for zero-crossing method

+++ src/rdf/bearing_calculator_base.rs
@@ -33,7 +33,7 @@ impl BearingCalculatorBase {
         let filter_group_delay = bandpass.group_delay_samples();
 
         Ok(Self {
-            agc: AutomaticGainControl::new(agc_config, sample_rate as u32),
+            agc: AutomaticGainControl::new(agc_config, sample_rate),
             bandpass,
             filter_group_delay,
             sample_counter: 0,
